feat: retro shell on all pages and full-width window
Move title bar, menu, toolbar, and Help dialog into RetroTop for every route.
Layout wraps non-home pages with RetroSiteNav plus main content slot; home
uses HomeDesktop for section tabs and scroll-spy status bar.

Replace Tailwind-heavy inner pages with retro utility classes; blog posts use
retro-prose. Remove per-page theme DOM scripts and old Navigation from layout.

Drop max-width on retro-window; rename body.retro-home to retro-site; widen
single-page main content (no 640px cap).
