+++ src/pages/blog/index.astro
@@ -8,103 +8,25 @@ const posts = (await getCollection('blog')).sort(
 ---
 
 <Layout title="Blog | Hammadh Arquil">
-	<div id="blog-page" class="container mt-8 sm:mt-12 px-4 sm:px-0 mx-auto">
-        <h1 class="text-3xl font-bold text-[var(--accent)] mb-8">Writing</h1>
-		<ul class="space-y-8 sm:space-y-10">
+	<div id="blog-page">
+		<h1 class="retro-page-h1">Writing</h1>
+		<ul class="retro-blog-list">
 			{
 				posts.map((post) => (
-					<li class="group">
-						<a href={`/blog/${post.slug}/`} class="block">
-                            <div class="flex flex-col sm:flex-row justify-between sm:items-baseline gap-1 sm:gap-2">
-							    <h4 class="text-lg sm:text-xl font-bold text-white group-hover:text-[var(--accent)] transition-colors leading-tight">{post.data.title}</h4>
-                                <p class="text-zinc-500 text-xs sm:text-sm font-mono shrink-0 mt-1 sm:mt-0">
-                                    {post.data.pubDate.toLocaleDateString('en-US', { month: 'short', day: 'numeric', year: 'numeric' })}
-                                </p>
-                            </div>
-							<p class="text-zinc-400 mt-2 text-sm sm:text-base leading-relaxed max-w-2xl">
-								{post.data.description}
+					<li class="retro-blog-item">
+						<a href={`/blog/${post.slug}/`}>
+							<h2 class="retro-blog-title">{post.data.title}</h2>
+							<p class="retro-blog-meta">
+								{post.data.pubDate.toLocaleDateString('en-US', { month: 'short', day: 'numeric', year: 'numeric' })}
 							</p>
-                            <div class="mt-3 text-sm text-[#bede64] font-medium opacity-0 -translate-x-2 group-hover:opacity-100 group-hover:translate-x-0 transition-all duration-300 flex items-center">
-                                Read Article
-                                <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="ml-1"><path d="M5 12h14"/><path d="m12 5 7 7-7 7"/></svg>
-                            </div>
+							<p class="retro-blog-desc">{post.data.description}</p>
+							<p class="retro-blog-read">Read article →</p>
 						</a>
 					</li>
 				))
 			}
 		</ul>
 
-        <div class="mt-[4em] pb-[3em]">
-             <p class="text-xs sm:text-sm font-normal text-zinc-500">
-                ©️ 2026 Hammadh Arquil. All Rights Reserved.
-            </p>
-        </div>
+		<p class="retro-footer" style="margin-top: 2rem;">© 2026 Hammadh Arquil. All rights reserved.</p>
 	</div>
 </Layout>
-
-<script>
-    import { isLightMode } from '../../stores/themeStore';
-    
-    isLightMode.subscribe(light => {
-        const page = document.getElementById('blog-page');
-        if (page) {
-            if (light) {
-                page.classList.add('light-mode');
-                 const title = page.querySelector('h1');
-                if (title) title.classList.remove('text-[#bede64]');
-                if (title) title.classList.add('text-[#1a1a1a]');
-
-                 const postTitles = page.querySelectorAll('h4');
-                postTitles.forEach(t => {
-                    t.classList.remove('text-white');
-                    t.classList.add('text-[#1a1a1a]');
-                });
-
-                 const postDescs = page.querySelectorAll('.text-zinc-400, .text-zinc-500');
-                postDescs.forEach(t => {
-                    t.classList.remove('text-zinc-400', 'text-zinc-500');
-                    t.classList.add('text-[#5a5a5a]');
-                });
-                
-                const readLinks = page.querySelectorAll('.text-[#bede64]');
-                 readLinks.forEach(t => {
-                     // Note: main title is also #bede64, but we handled it. 
-                     // This targets the "Read Article" link specifically if we want to change it, 
-                     // but #bede64 might look okay or we might want to darken it.
-                     // The title was handled above, this catches the list items.
-                     if (!t.classList.contains('text-3xl')) { 
-                         t.classList.remove('text-[#bede64]');
-                         t.classList.add('text-[#1a1a1a]'); 
-                     }
-                });
-
-
-            } else {
-                page.classList.remove('light-mode');
-                const title = page.querySelector('h1');
-                if (title) title.classList.add('text-[#bede64]');
-                if (title) title.classList.remove('text-[#1a1a1a]');
-
-                 const postTitles = page.querySelectorAll('h4');
-                postTitles.forEach(t => {
-                    t.classList.add('text-white');
-                    t.classList.remove('text-[#1a1a1a]');
-                });
-
-                 const postDescs = page.querySelectorAll('.text-[#5a5a5a]');
-                postDescs.forEach(t => {
-                    t.classList.add('text-zinc-400'); // using 400 for general
-                    t.classList.remove('text-[#5a5a5a]');
-                });
-
-                 const readLinks = page.querySelectorAll('.text-[#1a1a1a]');
-                 readLinks.forEach(t => {
-                     if (!t.classList.contains('text-3xl')) {
-                         t.classList.add('text-[#bede64]');
-                         t.classList.remove('text-[#1a1a1a]');
-                     }
-                });
-            }
-        }
-    });
-</script>
