+++ src/pages/contact.astro
@@ -3,98 +3,27 @@ import Layout from '../layouts/Layout.astro';
 ---
 
 <Layout title="Contact | Hammadh Arquil">
-    <div id="contact-page" class="container mt-8 sm:mt-12 h-[60vh] flex flex-col justify-center max-w-2xl">
-        <h1 class="text-4xl font-bold text-[#bede64] mb-6">Let's Connect</h1>
-        <p class="text-zinc-400 mb-8 text-lg leading-relaxed">
-            I'm currently looking for new opportunities to build exciting products. 
-            Whether you have a question, a project idea, or just want to say hi, I'll try my best to get back to you!
-        </p>
-        
-        <div class="space-y-4">
-             <a href="mailto:hammadharquil@gmail.com" class="block w-full p-4 border border-zinc-700 rounded-lg hover:border-[#bede64] hover:bg-zinc-800/50 transition-all group">
-                <h3 class="text-white font-medium group-hover:text-[#bede64] mb-1">Email</h3>
-                <p class="text-zinc-500 text-sm">hammadharquil@gmail.com</p>
-            </a>
-            
-             <div class="grid grid-cols-2 gap-4">
-                <a href="https://www.linkedin.com/in/hammadharquil/" target="_blank" class="p-4 border border-zinc-700 rounded-lg hover:border-[#bede64] hover:bg-zinc-800/50 transition-all">
-                    <span class="text-zinc-300 font-medium hover:text-[#bede64]">LinkedIn</span>
-                </a>
-                <a href="https://twitter.com/elitepsd" target="_blank" class="p-4 border border-zinc-700 rounded-lg hover:border-[#bede64] hover:bg-zinc-800/50 transition-all">
-                    <span class="text-zinc-300 font-medium hover:text-[#bede64]">Twitter / X</span>
-                </a>
-                <a href="https://github.com/hammvdh" target="_blank" class="p-4 border border-zinc-700 rounded-lg hover:border-[#bede64] hover:bg-zinc-800/50 transition-all">
-                    <span class="text-zinc-300 font-medium hover:text-[#bede64]">GitHub</span>
-                </a>
-                 <a href="https://instagram.com/hammvdh" target="_blank" class="p-4 border border-zinc-700 rounded-lg hover:border-[#bede64] hover:bg-zinc-800/50 transition-all">
-                    <span class="text-zinc-300 font-medium hover:text-[#bede64]">Instagram</span>
-                </a>
-             </div>
-        </div>
-    </div>
+	<div id="contact-page">
+		<h1 class="retro-page-h1">Let's connect</h1>
+		<p class="retro-contact-intro">
+			I'm currently looking for new opportunities to build exciting products. Whether you have a question, a project idea, or
+			just want to say hi, I'll try my best to get back to you!
+		</p>
+
+		<a href="mailto:hammadharquil@gmail.com" class="retro-contact-primary">
+			<h3>Email</h3>
+			<p>hammadharquil@gmail.com</p>
+		</a>
+
+		<div class="retro-contact-grid">
+			<a href="https://www.linkedin.com/in/hammadharquil/" target="_blank" rel="noopener noreferrer" class="retro-contact-card">
+				LinkedIn
+			</a>
+			<a href="https://twitter.com/elitepsd" target="_blank" rel="noopener noreferrer" class="retro-contact-card">Twitter / X</a>
+			<a href="https://github.com/hammvdh" target="_blank" rel="noopener noreferrer" class="retro-contact-card">GitHub</a>
+			<a href="https://instagram.com/hammvdh" target="_blank" rel="noopener noreferrer" class="retro-contact-card">Instagram</a>
+		</div>
+
+		<p class="retro-footer" style="margin-top: 2rem;">© 2026 Hammadh Arquil. All rights reserved.</p>
+	</div>
 </Layout>
-
-<script>
-    import { isLightMode } from '../stores/themeStore';
-    
-    isLightMode.subscribe(light => {
-        const page = document.getElementById('contact-page');
-        if (page) {
-            if (light) {
-                page.classList.add('light-mode');
-                 const title = page.querySelector('h1');
-                if (title) title.classList.remove('text-[#bede64]');
-                if (title) title.classList.add('text-[#1a1a1a]');
-                 const sub = page.querySelector('p');
-                if (sub) sub.classList.remove('text-zinc-400');
-                if (sub) sub.classList.add('text-[#4a4a4a]');
-
-                const cards = page.querySelectorAll('.border-zinc-700');
-                cards.forEach(c => {
-                    c.classList.remove('border-zinc-700', 'hover:bg-zinc-800/50');
-                    c.classList.add('border-zinc-200', 'hover:bg-zinc-100');
-                });
-
-                const whiteText = page.querySelectorAll('.text-white');
-                whiteText.forEach(t => {
-                    t.classList.remove('text-white');
-                    t.classList.add('text-[#1a1a1a]');
-                });
-
-                  const zincText = page.querySelectorAll('.text-zinc-300, .text-zinc-500');
-                zincText.forEach(t => {
-                    t.classList.remove('text-zinc-300', 'text-zinc-500');
-                    t.classList.add('text-[#5a5a5a]');
-                });
-
-
-            } else {
-                page.classList.remove('light-mode');
-                const title = page.querySelector('h1');
-                if (title) title.classList.add('text-[#bede64]');
-                if (title) title.classList.remove('text-[#1a1a1a]');
-                 const sub = page.querySelector('p');
-                if (sub) sub.classList.add('text-zinc-400');
-                if (sub) sub.classList.remove('text-[#4a4a4a]');
-
-                 const cards = page.querySelectorAll('.border-zinc-200');
-                cards.forEach(c => {
-                    c.classList.add('border-zinc-700', 'hover:bg-zinc-800/50');
-                    c.classList.remove('border-zinc-200', 'hover:bg-zinc-100');
-                });
-
-                 const blackText = page.querySelectorAll('.text-[#1a1a1a]');
-                blackText.forEach(t => {
-                    t.classList.add('text-white');
-                    t.classList.remove('text-[#1a1a1a]');
-                });
-
-                  const zincText = page.querySelectorAll('.text-[#5a5a5a]');
-                zincText.forEach(t => {
-                    t.classList.add('text-zinc-300'); // simplistic reset
-                    t.classList.remove('text-[#5a5a5a]');
-                });
-            }
-        }
-    });
-</script>
