+++ src/pages/index.astro
@@ -1,9 +1,6 @@
 ---
 import Layout from '../layouts/Layout.astro';
-import PortfolioShell from '../components/PortfolioShell';
 ---
 
-<Layout title="Hammadh Arquil | Software Engineer" hideNavigation={true}>
-	<PortfolioShell client:load />
-</Layout>
+<Layout title="Hammadh Arquil | Software Engineer" />
